+++ src/adapter/web-standard/handler.ts
@@ -16,6 +16,9 @@ import { ElysiaCustomStatusResponse } from '../../error'
 
 import type { Context } from '../../context'
 import type { AnyLocalHook } from '../../types'
+import { isBun } from '../../universal/utils'
+
+const isNotBun = !isBun
 
 export const mapResponse = (
 	response: unknown,
@@ -27,6 +30,9 @@ export const mapResponse = (
 
 		switch (response?.constructor?.name) {
 			case 'String':
+				if (isNotBun && !set.headers['content-type'])
+					set.headers['content-type'] = 'text/plain'
+
 				return new Response(response as string, set as any)
 
 			case 'Array':
@@ -195,6 +201,9 @@ export const mapEarlyResponse = (
 
 		switch (response?.constructor?.name) {
 			case 'String':
+				if (isNotBun && !set.headers['content-type'])
+					set.headers['content-type'] = 'text/plain'
+
 				return new Response(response as string, set as any)
 
 			case 'Array':
@@ -330,6 +339,9 @@ export const mapEarlyResponse = (
 	} else
 		switch (response?.constructor?.name) {
 			case 'String':
+				if (isNotBun && !set.headers['content-type'])
+					set.headers['content-type'] = 'text/plain'
+
 				return new Response(response as string)
 
 			case 'Array':
@@ -478,7 +490,13 @@ export const mapCompactResponse = (
 ): Response => {
 	switch (response?.constructor?.name) {
 		case 'String':
-			return new Response(response as string)
+			if (isNotBun) return new Response(response as string)
+
+			return new Response(response as string, {
+				headers: {
+					'Content-Type': 'text/plain'
+				}
+			})
 
 		case 'Object':
 		case 'Array':